ADEN-12179 - Added pipeline-approach initialization to sports platform, added PPID to sports platform
