+++ platforms/sports/modes/ads/sports-ads.mode.ts
@@ -1,33 +1,39 @@
-import { startAdEngine, wadRunner } from '@platforms/shared';
+import { gptSetup, wadRunner } from '@platforms/shared';
 import {
 	audigent,
 	bidders,
+	communicationService,
 	confiant,
 	DiProcess,
 	durationMedia,
+	eventsRepository,
 	iasPublisherOptimization,
+	PartnerPipeline,
+	userIdentity,
 } from '@wikia/ad-engine';
 import { Injectable } from '@wikia/dependency-injection';
 
 @Injectable()
 export class SportsAdsMode implements DiProcess {
+	constructor(private pipeline: PartnerPipeline) {}
+
 	execute(): void {
-		const inhibitors = this.callExternals();
-
-		startAdEngine(inhibitors);
-	}
-
-	private callExternals(): Promise<any>[] {
-		const inhibitors: Promise<any>[] = [];
-
-		inhibitors.push(bidders.call());
-		inhibitors.push(wadRunner.call());
-
-		audigent.call();
-		iasPublisherOptimization.call();
-		confiant.call();
-		durationMedia.call();
-
-		return inhibitors;
+		this.pipeline
+			.add(
+				userIdentity,
+				bidders,
+				wadRunner,
+				audigent,
+				iasPublisherOptimization,
+				confiant,
+				durationMedia,
+				gptSetup.setOptions({
+					dependencies: [userIdentity.initialized],
+				}),
+			)
+			.execute()
+			.then(() => {
+				communicationService.emit(eventsRepository.AD_ENGINE_PARTNERS_READY);
+			});
 	}
 }
