+++ platforms/sports/sports-platform.ts
@@ -22,6 +22,7 @@ import {
 import { Injectable } from '@wikia/dependency-injection';
 import { getBasicContext } from './ad-context';
 import { SportsAdsMode } from './modes/ads/sports-ads.mode';
+import { SportsAdsDeprecatedMode } from './modes/ads/sports-ads-deprecated.mode';
 import { SportsA9ConfigSetup } from './setup/context/a9/sports-a9-config.setup';
 import { SportsPrebidConfigSetup } from './setup/context/prebid/sports-prebid-config.setup';
 import { SportsSlotsContextSetup } from './setup/context/slots/sports-slots-context.setup';
@@ -64,7 +65,10 @@ export class SportsPlatform {
 		// Run
 		this.pipeline.add(
 			conditional(() => this.noAdsDetector.isAdsMode(), {
-				yes: SportsAdsMode,
+				yes: conditional(() => context.get('options.adsInitializeV2'), {
+					yes: SportsAdsMode,
+					no: SportsAdsDeprecatedMode,
+				}),
 			}),
 		);
